% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_attendance}
\alias{dese_attendance}
\title{Attendance and Retention Report}
\usage{
dese_attendance(year, mode = "school")
}
\arguments{
\item{year}{a character string with the year (e.g., school year 2015-2016 is "2016")}

\item{mode}{district or school}
}
\value{
data.frame
}
\description{
Get the \href{http://profiles.doe.mass.edu/state_report/indicators.aspx}{Attendance and Retention} Report from Massachusetts DESE
}
\examples{
dese_attendance("2016")
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

