% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_attrition}
\alias{dese_attrition}
\title{Attrition}
\usage{
dese_attrition(year, mode = "SCHOOL", group = "ALL")
}
\arguments{
\item{year}{the year of the report}

\item{mode}{district or school}

\item{group}{get the report for a specific student group. Defaults to all students.}
}
\description{
Get the \href{http://profiles.doe.mass.edu/state_report/attrition.aspx}{Attrition}
Report from Massachusetts DESE website.
Attrition is students who left the school over the summer.
}
\examples{
dese_attrition("2016")
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

