% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_enrollmentbyracegender}
\alias{dese_enrollmentbyracegender}
\title{Enrollment by Race and Gender}
\usage{
dese_enrollmentbyracegender(year, mode = "school")
}
\arguments{
\item{year}{a character string with the numeric year (e.g., school year 2015-2016 is "2016")}

\item{mode}{district or school}
}
\value{
data.frame
}
\description{
Get the \href{http://profiles.doe.mass.edu/state_report/enrollmentbyracegender.aspx}{Enrollemnt by Race/Gender} Report from Massachusetts DESE
}
\examples{
dese_enrollmentbyracegender("2016")
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

