% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_get}
\alias{dese_get}
\title{Get data from Massachusetts DESE}
\usage{
dese_get(report, query = NULL)
}
\arguments{
\item{report}{the name of the report you are getting}

\item{query}{any additional query needed to complete the url}
}
\value{
data.frame
}
\description{
Get public school information from Massachusetts \href{http://profiles.doe.mass.edu/state_report/}{State Reports}.
}
\details{
A wrapper function that execute a GET to http://profiles.doe.mass.edu/state_report/
and parses the results.
}
\examples{
dese_get('enrollmentbyracegender', query=list('mode'='school', 'year'='2016'))
}
\seealso{
\code{\link{GET}} and \code{\link{rvest}} which this function wraps
}

