% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_mcas}
\alias{dese_mcas}
\title{MCAS Achievement Results}
\usage{
dese_mcas(year, SchoolType = "All", studentGroup = "AL", grade = "AL",
  mode = "school")
}
\arguments{
\item{year}{a character string with the numeric year (e.g., school year 2015-2016 is "2016")}

\item{SchoolType}{All, Elementary, Middle, etc}

\item{studentGroup}{'AL' for all,}

\item{grade}{'AL' for all, 10, 08, etc}

\item{mode}{district or school}
}
\value{
data.frame
}
\description{
Get \href{http://profiles.doe.mass.edu/state_report/mcas.aspx}{MCAS} Reports from Massachusetts DESE
}
\examples{
dese_mcas("2014")
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

