% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_mobilityrates}
\alias{dese_mobilityrates}
\title{Mobility Rates}
\usage{
dese_mobilityrates(year, mode = "SCHOOL", group = "ALL")
}
\arguments{
\item{year}{the year of the report}

\item{mode}{district or school}

\item{group}{get the report for a specific student group. Defaults to all students.}
}
\description{
Get the \href{http://profiles.doe.mass.edu/state_report/mobilityrates.aspx}{Mobility Rates}
Report from Massachusetts DESE website.
Mobility rates has various metrics on how many students move through the school in a year
}
\examples{
dese_mobilityrates(2016)
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

