% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_multiget.R
\name{dese_multiget}
\alias{dese_multiget}
\title{dese_multiget}
\usage{
dese_multiget(report, ...)
}
\arguments{
\item{report}{Quoted name of the report where the name is an madese function with form dese_[report]}

\item{year}{One or more years to get (as a list if multiple)}

\item{mode}{One or more modes to get (usually SCHOOL)}

\item{group}{One or more student sub-groups to get}
}
\description{
Note: Not working.
Get multiple years, groups and/or modes in a single call.
The result is a named list of dataframes where the names have the format
"[report]_[mode]_[group]_[year]"
}
\examples{
mode <- "SCHOOL"
years <- 2015:2016
groups <- "ALL"
dese_multiget('enrollmentbyracegender', modes=mode, groups=groups, years=years))
}

