% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_assessments.R
\name{dese_parcc}
\alias{dese_parcc}
\title{dese_parcc}
\usage{
dese_parcc(year = "All", mode = "All", group = "All",
  schooltype = "All", testtype = "All")
}
\arguments{
\item{year}{a character string with the numeric year (e.g., school year 2015-2016 is "2016")}

\item{mode}{district or school}

\item{group}{Student group}

\item{schooltype}{middle school, elementary school, etc}

\item{testtype}{The grade / subject of the test}
}
\description{
Get the PARCC Report from Massachusetts DESE
}

