% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_selectedpopulations}
\alias{dese_selectedpopulations}
\title{Enrollment by Selected Populations}
\usage{
dese_selectedpopulations(year, mode = "school")
}
\arguments{
\item{year}{a character string with the numeric year (e.g., school year 2015-2016 is "2016")}

\item{mode}{district or school}
}
\value{
data.frame
}
\description{
Get the \href{http://profiles.doe.mass.edu/state_report/selectedpopulations.aspx}{Enrollemnt by Selected Populations} Report from Massachusetts DESE 
(first language not english, english language learners, students with disabilities, etc)
}
\examples{
dese_selectedpopulations("2016")
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

