% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dese_get.R
\name{dese_ssdr_days_missed}
\alias{dese_ssdr}
\alias{dese_ssdr_days_missed}
\title{dese_ssdr_days_missed}
\usage{
dese_ssdr_days_missed(year, mode = "SCHOOL", group = "ALL",
  offense = "All")

dese_ssdr(year, mode = "SCHOOL", group = "ALL", offense = "All")
}
\arguments{
\item{year}{the year of the report}

\item{mode}{district or school}

\item{group}{get the report for a specific student group. Defaults to all students. Can be
- ALL = All students
- SPED = Students with disabilities
- ECODIS = Economically disadvantaged students
- F = Female students
- M = Male students
- AI = American Indian or Alaskan Native students
- AS = Asian students
- BL = Black / African American students
- HS = Hispanic / Latino students
- MR = Multiracial, non-hispanic / Latino students
- HP = Native Hawaiian / Pacific Islander students
- WH = White students}

\item{offense}{Defaults to all offenses.}
}
\description{
Get SSDR-based reports from Massachusetts DESE website.
}
\seealso{
\code{\link{dese_get}} which this function wraps
}

